¥¤¥󥹥ȡ¼¥ë»þ¤˥·¥ó¥ܥê¥å¯¥ê¥󥯤ϥХå¯¥¢¥åפ·¤ʤ¤¤褦¤ˤ·¤¿

+++ install.sh
@@ -3,36 +3,36 @@ git clone https://github.com/landcat222/dotfiles.git ~/.dotfiles
 mkdir ~/.backup
 
 # zsh files backup
-if [ -e ~/.zshenv ];then
+if [ -e ~/.zshenv -a ! -L ~/.zshenv ];then
   mv -b ~/.zshenv ~/.backup/zshenv
 fi
 
-if [ -e ~/.zshrc ];then
+if [ -e ~/.zshrc -a ! -L ~/.zshrc ];then
   mv -b ~/.zshrc ~/.backup/zshrc
 fi
 
-if [ -e ~/.zprofile ];then
+if [ -e ~/.zprofile -a ! -L ~/.zprofile ];then
   mv -b ~/.zprofile ~/.backup/zprofile
 fi
 
-if [ -e ~/.zlogin ];then
+if [ -e ~/.zlogin -a ! -L ~/.zlogin ];then
   mv -b ~/.zlogin ~/.backup/zlogin
 fi
 
-if [ -e ~/.zlogout ];then
+if [ -e ~/.zlogout -a ! -L ~/.zlogout ];then
   mv -b ~/.zlogout ~/.backup/zlogout
 fi
 
-if [ -e ~/.zsh ];then
+if [ -e ~/.zsh -a ! -L ~/.zsh ];then
   mv -b ~/.zsh ~/.backup/zsh
 fi
 
 # vim files backup
-if [ -e ~/.vimrc ];then
+if [ -e ~/.vimrc -a ! -L ~/.vimrc ];then
   mv -b ~/.vimrc ~/.backup/vimrc
 fi
 
-if [ -e ~/.vim ];then
+if [ -e ~/.vim -a ! -L ~/.vim ];then
   mv -b ~/.vim ~/.backup/vim
 fi
